feat: Disable Watchdogs when USB-Serial-JTAG is used
Watchdog needs to be disabled in order to prevent reset in the middle of
a communication. By default watchdog is disabled after chip reset,
but second stage bootloader may enable it again and if application does
not disable it, it will reset the device, because USB-Serial/JTAG does
only software reset.